put recency/kind in metrics-util + reorganize prometheus exporter
