+++ metrics-util/src/lib.rs
@@ -11,7 +11,7 @@ pub use bucket::AtomicBucket;
 #[cfg(feature = "std")]
 mod debugging;
 #[cfg(feature = "std")]
-pub use debugging::{DebugValue, DebuggingRecorder, MetricKind, Snapshotter};
+pub use debugging::{DebugValue, DebuggingRecorder, Snapshotter};
 
 #[cfg(feature = "std")]
 mod handle;
@@ -29,7 +29,15 @@ pub use registry::{Generation, Registry};
 mod key;
 pub use key::CompositeKey;
 
+mod kind;
+pub use kind::MetricKind;
+
 mod histogram;
 pub use histogram::Histogram;
 
 pub mod layers;
+
+#[cfg(feature = "std")]
+mod recency;
+#[cfg(feature = "std")]
+pub use recency::Recency;
